+++ src/main/java/compression/CompressionModelFactory.java
@@ -29,7 +29,7 @@ public class CompressionModelFactory {
 
     public static List<ValueCompressionModel> getValueCompressionModels() {
         List<ValueCompressionModelType> valueModelTypes =  config.getValueModels();
-        final double valueModelErrorBound = config.getValueModelErrorBound();
+        final float valueModelErrorBound = config.getValueModelErrorBound();
 
         return getCompressionModels(valueModelTypes, (modelType) -> CompressionModelFactory.getValueCompressionModelByType(modelType, valueModelErrorBound));
     }
@@ -42,7 +42,7 @@ public class CompressionModelFactory {
     }
 
 
-    private static ValueCompressionModel getValueCompressionModelByType(ValueCompressionModelType valueCompressionModelType, double errorBound) {
+    private static ValueCompressionModel getValueCompressionModelByType(ValueCompressionModelType valueCompressionModelType, float errorBound) {
         switch (valueCompressionModelType) {
             case PMC_MEAN:
                 return new PMCMeanValueCompressionModel(errorBound);
