+++ src/test/java/segmentgenerator/SegmentGeneratorTest.java
@@ -37,7 +37,7 @@ class SegmentGeneratorTest {
         allDataPoints.add(lastDataPoint);
         Segment expectedSegment = new Segment(1, 1, 6, ValueCompressionModelType.PMC_MEAN.ordinal(), constructValueDataBlob(allDataPoints), TimeStampCompressionModelType.REGULAR.ordinal(), constructTimestampDataBlob(allDataPoints));
 
-        double errorBound = 0;
+        float errorBound = 0;
         TestCompressionModelManagerRegularPMCMean testCompressionModelManagerRegularPMCMean = new TestCompressionModelManagerRegularPMCMean(List.of(new PMCMeanValueCompressionModel(errorBound)), List.of(new RegularTimeStampCompressionModel(errorBound)));
 
         SegmentGenerator segmentGenerator = new SegmentGenerator(testCompressionModelManagerRegularPMCMean, 1);
